OmniAuth mock functional in spec
