+++ app/controllers/sessions_controller.rb
@@ -4,7 +4,8 @@ class SessionsController < ApplicationController
     client_id = ENV["GOOGLE_CLIENT_ID"]
     client_secret = ENV["GOOGLE_CLIENT_SECRET"]
     code = params[:code]
-    #response = Faraday.post
-    binding.pry
+    user = User.update_or_create(request.env["omniauth.auth"])
+    session[:id] = user.id
+    redirect_to root_path
   end
 end
