+++ app/models/user.rb
@@ -15,5 +15,5 @@ class User < ApplicationRecord
     user.save!
     user
   end
-  
+
 end
