+++ spec/features/welcome/log_in_spec.rb
@@ -1,0 +1,46 @@
+require 'rails_helper'
+
+RSpec.feature "user logs in" do
+  scenario "using google oauth2" do
+    stub_omniauth
+    visit root_path
+    expect(page).to have_link("Sign in with Google")
+    click_link "Sign in with Google"
+
+    expect(page).to have_content("Lito White")
+    expect(page).to have_link("Logout")
+  end
+
+  def stub_omniauth
+      OmniAuth.config.test_mode = true
+      OmniAuth.config.mock_auth[:google_oauth2] = OmniAuth::AuthHash.new({
+        provider: "google",
+        uid: "12345678910",
+        info: {
+          email: "lito@gmail.com",
+          first_name: "Lito",
+          last_name: "White"
+        },
+        credentials: {
+          token: "abcdefg12345",
+          refresh_token: "12345abcdefg",
+          expires_at: DateTime.now,
+        }
+      })
+      #
+      # OmniAuth.config.add_mock(:google => {
+      #   provider: "google",
+      #   uid: "12345678910",
+      #   info: {
+      #     email: "lito@gmail.com",
+      #     first_name: "Lito",
+      #     last_name: "White"
+      #   },
+      #   credentials: {
+      #     token: "abcdefg12345",
+      #     refresh_token: "12345abcdefg",
+      #     expires_at: DateTime.now,
+      #   }
+      # })
+    end
+end
